requirements.txt
# Web framework
fastapi>=0.110.0
uvicorn[standard]>=0.29.0

# PDF processing
pdf2image>=1.17.0
Pillow>=10.3.0

# HTTP client (OpenRouter calls)
httpx>=0.27.0

# Data validation
pydantic>=2.7.0
pydantic-settings>=2.2.0

# Async task handling
anyio>=4.3.0

# Unique token generation
python-ulid>=2.0.0

# Testing
pytest>=8.2.0
pytest-asyncio>=0.23.0
httpx  # also used as AsyncClient in tests
